Update Resource machine to v0.2 (#43)
* update SimpleHelloWorld to v0.2

* counter wip

* fix Counter Simple and Unique

* kudos wip

* kudos compiles

* update counters to applib rc12

* use applib rc13

* update HelloWorld

* update anoma

* Fix run-tests.sh so error codes are propagated

`set -e` is required so that the error code from the parallel command is
propagated to the parent shell.

`set -o pipefail` is required in the parallel command so that the error
code from `./all-tests.sh` is propagated to the parent shell. Otherwise
this error code is masked by `| tee`.

* refac: Bump versions in HelloWorld README

* update versions of dependencies

* remove Extra package

---------

Co-authored-by: Paul Cadman <git@paulcadman.dev>
Co-authored-by: mauricewbr <39859692+mauricewbr@users.noreply.github.com>